feat(rooms): add amenity management, update db schema & auto-fill form

+++ backend/src/main/java/com/hotel/modules/rooms/controller/RoomTypeController.java
@@ -28,7 +28,7 @@ public class RoomTypeController {
     }
 
     @GetMapping("/hotel/{hotelId}")
-    public ResponseEntity<List<RoomTypeResponse>> getByHotelId(@PathVariable Integer hotelId) {
+    public ResponseEntity<List<RoomTypeResponse>> getByHotelId(@PathVariable Long hotelId) {
         return ResponseEntity.ok(roomTypeService.getByHotelId(hotelId));
     }
